move test classes to their packages
