+++ eo-maven-plugin/src/test/java/org/eolang/maven/AssembleMojoTest.java
@@ -25,7 +25,6 @@ package org.eolang.maven;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import org.cactoos.func.UncheckedFunc;
 import org.cactoos.io.InputOf;
 import org.cactoos.set.SetOf;
 import org.eolang.maven.util.Home;
@@ -73,9 +72,7 @@ final class AssembleMojoTest {
             .with("ignoreTransitive", true)
             .with(
                 "objectionary",
-                new OyLambda(
-                    new UncheckedFunc<>(s -> new InputOf("[] > sprintf\n"))
-                )
+                new OyFake(s -> new InputOf("[] > sprintf\n"))
             )
             .execute();
         MatcherAssert.assertThat(
@@ -134,9 +131,7 @@ final class AssembleMojoTest {
             .with("ignoreTransitive", true)
             .with(
                 "objectionary",
-                new OyLambda(
-                    new UncheckedFunc<>(s -> new InputOf("[] > sprintf\n"))
-                )
+                new OyFake(s -> new InputOf("[] > sprintf\n"))
             )
             .execute();
         MatcherAssert.assertThat(
