+++ eo-maven-plugin/src/test/java/org/eolang/maven/CleanMojoTest.java
@@ -27,7 +27,6 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import org.cactoos.func.UncheckedFunc;
 import org.cactoos.io.InputOf;
 import org.cactoos.set.SetOf;
 import org.eolang.maven.util.Home;
@@ -92,9 +91,7 @@ class CleanMojoTest {
             .with("ignoreTransitive", true)
             .with(
                 "objectionary",
-                new OyLambda(
-                    new UncheckedFunc<>(s -> new InputOf("[] > sprintf\n"))
-                )
+                new OyFake(s -> new InputOf("[] > sprintf\n"))
             )
             .execute();
         new Moja<>(CleanMojo.class)
