+++ eo-maven-plugin/src/test/java/org/eolang/maven/PullMojoTest.java
@@ -28,7 +28,6 @@ import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.LinkedList;
-import org.cactoos.func.UncheckedFunc;
 import org.cactoos.io.InputOf;
 import org.cactoos.io.ResourceOf;
 import org.eolang.maven.objectionary.Objectionary;
@@ -136,9 +135,7 @@ final class PullMojoTest {
      * @return Dummy Objectionary.
      */
     private Objectionary dummy() {
-        return new OyLambda(
-            new UncheckedFunc<>(s -> new InputOf("[] > hello\n"))
-        );
+        return new OyFake(s -> new InputOf("[] > hello\n"));
     }
 
 }
