+++ eo-maven-plugin/src/test/java/org/eolang/maven/SkipTest.java
@@ -27,7 +27,6 @@ package org.eolang.maven;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import org.cactoos.func.UncheckedFunc;
 import org.cactoos.io.InputOf;
 import org.eolang.maven.util.Home;
 import org.eolang.maven.util.Online;
@@ -116,9 +115,7 @@ class SkipTest {
             .with("skip", skip)
             .with(
                 "objectionary",
-                new OyLambda(
-                    new UncheckedFunc<>(s -> new InputOf("[] > hello\n"))
-                )
+                new OyFake(s -> new InputOf("[] > hello\n"))
             )
             .execute();
     }
