+++ eo-maven-plugin/src/test/java/org/eolang/maven/objectionary/OyCachingTest.java
@@ -21,33 +21,42 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.maven;
+package org.eolang.maven.objectionary;
 
 import java.nio.file.Path;
-import java.nio.file.Paths;
-import org.cactoos.list.ListOf;
-import org.eolang.maven.util.Home;
-import org.eolang.maven.util.Walk;
+import org.cactoos.io.InputOf;
+import org.cactoos.text.TextOf;
+import org.eolang.maven.OyFake;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 /**
- * Test case for {@link Walk}.
+ * Test for {@link OyCaching}.
  *
- * @since 0.11
+ * @since 1.0
  */
-final class WalkTest {
+final class OyCachingTest {
 
     @Test
-    void findsFiles(@TempDir final Path temp) throws Exception {
-        new Home(temp).save("", Paths.get("foo/hello/0.1/EObar/x.bin"));
-        new Home(temp).save("", Paths.get("EOxxx/bar"));
+    void putsObjectToLocalCache(@TempDir final Path path) throws Exception {
+        final String content = "[] > main\n";
         MatcherAssert.assertThat(
-            new Walk(temp).includes(new ListOf<>("EO**/*")),
-            Matchers.iterableWithSize(1)
+            new TextOf(
+                new OyCaching(
+                    "master",
+                    path,
+                    new OyFake(s -> new InputOf(content))
+                ).get("org.example.main")
+            ).asString(),
+            Matchers.is(content)
+        );
+        Assertions.assertTrue(
+            path.resolve("pulled/master/org/example/main.eo")
+                .toFile()
+                .exists()
         );
     }
-
 }
